sources/oauth: add entra ID source and move logic over
