+++ authentik/sources/oauth/models.py
@@ -232,12 +232,23 @@ class GoogleOAuthSource(CreatableType, OAuthSource):
 
 
 class AzureADOAuthSource(CreatableType, OAuthSource):
-    """Social Login using Azure AD."""
+    """(Deprecated) Social Login using Azure AD."""
 
     class Meta:
         abstract = True
         verbose_name = _("Azure AD OAuth Source")
         verbose_name_plural = _("Azure AD OAuth Sources")
+
+
+# TODO: When removing this, add a migration for OAuthSource that sets
+# provider_type to `entraid` if it is currently `azuread`
+class EntraIDOAuthSource(CreatableType, OAuthSource):
+    """Social Login using Entra ID."""
+
+    class Meta:
+        abstract = True
+        verbose_name = _("Entra ID OAuth Source")
+        verbose_name_plural = _("Entra ID OAuth Sources")
 
 
 class OpenIDConnectOAuthSource(CreatableType, OAuthSource):
